requirement.txt
fastapi==0.115.8
uvicorn==0.34.0
websockets==11.0.3
httptools
pydantic==2.10.6
pyyaml==6.0.2
requests==2.32.3
pandas==2.2.3
tqdm==4.67.1
jsonschema==4.23.0
tenacity==9.0.0
matplotlib